*.c: use is_missing_file_error() where ENOENT and ENOTDIR are checked
These places were found using command:

  $ git grep -e ENOENT --and -e ENOTDIR

from commit c7054209d6 (treewide: use is_missing_file_error() where
ENOENT and ENOTDIR are checked, 2017-05-30).  There are codepaths that
only check ENOENT, and it is possible that some of them should be
checking both.  Updating them is kept out of this step deliberately, as
we do not want to change behaviour in this step.

Commit-message-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Andrei Rybak <rybak.a.v@gmail.com>